Drop support for React 0.13
