+++ package.json
@@ -50,9 +50,9 @@
     "jsdom": "~5.4.3",
     "mocha": "^2.2.5",
     "mocha-jsdom": "~0.4.0",
-    "react": "^0.14.0-rc1",
-    "react-addons-test-utils": "^0.14.0-rc1",
-    "react-dom": "^0.14.0-rc1",
+    "react": "^0.14.0",
+    "react-addons-test-utils": "^0.14.0",
+    "react-dom": "^0.14.0",
     "redux": "^3.0.0",
     "rimraf": "^2.3.4",
     "webpack": "^1.11.0"
@@ -62,6 +62,7 @@
     "invariant": "^2.0.0"
   },
   "peerDependencies": {
+    "react": "^0.14.0",
     "redux": "^2.0.0 || ^3.0.0"
   }
 }
