--- src/components/createAll.js
@@ -1,9 +1,0 @@
-import createProvider from './createProvider';
-import createConnect from './createConnect';
-
-export default function createAll(React) {
-  const Provider = createProvider(React);
-  const connect = createConnect(React);
-
-  return { Provider, connect };
-}
