--- src/native.js
@@ -1,4 +1,0 @@
-import React from 'react-native';
-import createAll from './components/createAll';
-
-export const { Provider, connect } = createAll(React);
