+++ src/utils/storeShape.js
@@ -1,0 +1,7 @@
+import { PropTypes } from 'react';
+
+export default PropTypes.shape({
+  subscribe: PropTypes.func.isRequired,
+  dispatch: PropTypes.func.isRequired,
+  getState: PropTypes.func.isRequired
+});
