+++ webpack.config.base.js
@@ -19,7 +19,6 @@ var reduxExternal = {
 module.exports = {
   externals: {
     'react': reactExternal,
-    'react-native': reactExternal,
     'redux': reduxExternal
   },
   module: {
